refactor app page
